Add a toggle for screen dimming when selecting a region.
I am not sure about others, but I have around a 2 second delay while
attempting to capture a screen region. Coming from puush (a similar
application I had a hand in writing) which did not draw screen overlays
but instead worked with a live image, this feels very cumbersome.

By disabling this option ShareX will not longer render the two darkness
levels before displaying the capture form. I have left the default as
enabled as to not change the default user experience.

As a further suggestion, it may be feasible to do away with the
"lightBackgroundBrush" layer, instead replacing it with the original
image. This would halve the processing time and still provide the a
level of contrast.