Fix ShareX/ShareX#3221 by throttling Invalidate calls
Invalidate should not be called at the end of every OnPaint, since it will immediately trigger another OnPaint, creating an "infinite" loop. Instead, Invalidate should be called only when we know the screen has changed in such a way that it needs to be repainted.

This commit changes the Invalidate calls to occur on an interval timer, at a maximum of once every 16ms (~60fps). While this is still high, the validation only occurs when relevant user input is detected (e.g. mouse movements, clicks, keypresses). After a 100ms timeout, the Invalidation timer is throttled back to a 250ms interval (~4fps) as a failsafe, which allows the borderDotPen animation to continue even when no input is detected. The result is smooth UI when the user is interacting with the app and efficient CPU usage when the user is idling. Resource usage is negligible at 4fps and still manageable in short bursts of user interaction at 60fps without chewing up a whole core.

Additionally, the old Pause and Resume actions still function as they did before, so no paints occur when the editor/region capture form is not in the foreground.

This commit will probably have a positive impact on the following issues too:
- ShareX/ShareX#3555
- ShareX/ShareX#3386
- ShareX/ShareX#3326
- ShareX/ShareX#3132