Ensure proper saving of settings on file shares
On Windows file shares (DFS) with restricted WRITE_DAC permission, the method File.Replace results in a System.UnauthorizedAccessException, even when the user has full permissions to interact with the file and folder. This configuration can commonly be found in business environments.

By setting the flag ignoreMetadataErrors to true, errors related to attributes and ACL are ignored. This results in the settings being saved correctly. Normal errors during file operation are not suppressed with this change, they should fail as usual.